refactor(builtins): split uniprop.rs into cohesive submodules (§7-8)
Split the 1187-line builtins/uniprop.rs (over the 500-line limit) into a
facade plus four themed submodules under src/builtins/uniprop/. All files
are under 500 lines.

- uniprop.rs (9)             — facade: mod decls + pub(crate) re-exports
- binary_props.rs (108)      — binary property helpers: check_binary_property,
                               try_binary_property, NFKC/composition helpers
- char_props.rs (429)        — per-char string props: numeric, case mapping,
                               bidi class, decomp type, east asian width, hangul
- text_seg.rs (212)          — text segmentation: grapheme cluster break,
                               joining type/group, sentence/word/line break
- lookup.rs (449)            — main entry points + normalization QC props +
                               emoji + unimatch functions + bidi mirroring

Pure relocation. All 16 external crate::builtins::uniprop:: call sites
(in builtins/functions/, builtins/methods_0arg/, builtins/methods_narg/,
runtime/unicode.rs) are preserved without changes — the facade re-exports
unicode_property_value, unicode_property_value_for_codepoint, unimatch,
unimatch_for_codepoint, and unicode_numeric_type at their original
pub(crate) visibility. Body refs to parent-level siblings rewritten from
super::unicode::*/super::methods_0arg::*/super::uniprop_tables::* to
absolute crate::builtins::* paths. Cross-submodule refs use super:: within
the uniprop/ tree. The fn multiset is byte-identical.
